requirements.txt
# borsh-construct
# construct
# construct-typing
construct
solders
websockets
aiohttp
python-dotenv
